[AI-2416] Revert "Implement automatic exposition format detection (#14445)"
This reverts commit 7d8dc6d28ec16239b32b2549daa27ff9e4289cd8.
