+++ datadog_checks_base/datadog_checks/base/checks/openmetrics/v2/scraper.py
@@ -9,9 +9,8 @@ from itertools import chain
 from math import isinf, isnan
 from typing import List  # noqa: F401
 
-from prometheus_client.openmetrics.parser import text_fd_to_metric_families as parse_openmetrics
-from prometheus_client.parser import text_fd_to_metric_families as parse_prometheus
-from requests.exceptions import ConnectionError
+from prometheus_client.openmetrics.parser import text_fd_to_metric_families as parse_metric_families_strict
+from prometheus_client.parser import text_fd_to_metric_families as parse_metric_families
 
 from ....config import is_affirmative
 from ....constants import ServiceCheck
@@ -219,17 +218,14 @@ class OpenMetricsScraper:
 
         self.http = RequestsWrapper(config, self.check.init_config, self.check.HTTP_CONFIG_REMAPPER, self.check.log)
 
-        self._content_type = ''
-        self._use_latest_spec = is_affirmative(config.get('use_latest_spec', False))
-        # Accept headers are taken from:
-        # https://github.com/prometheus/prometheus/blob/v2.43.0/scrape/scrape.go#L787
-        if self._use_latest_spec:
-            accept_header = 'application/openmetrics-text;version=1.0.0,application/openmetrics-text;version=0.0.1'
+        # Decide how strictly we will adhere to the latest version of the specification
+        if is_affirmative(config.get('use_latest_spec', False)):
+            self.parse_metric_families = parse_metric_families_strict
+            # https://github.com/prometheus/client_python/blob/v0.9.0/prometheus_client/openmetrics/exposition.py#L7
+            accept_header = 'application/openmetrics-text; version=0.0.1; charset=utf-8'
         else:
-            accept_header = (
-                'application/openmetrics-text;version=1.0.0,application/openmetrics-text;version=0.0.1;q=0.75,'
-                'text/plain;version=0.0.4;q=0.5,*/*;q=0.1'
-            )
+            self.parse_metric_families = parse_metric_families
+            accept_header = 'text/plain'
 
         # Request the appropriate exposition format
         if self.http.options['headers'].get('Accept') == '*/*':
@@ -284,15 +280,6 @@ class OpenMetricsScraper:
         if self.raw_line_filter is not None:
             line_streamer = self.filter_connection_lines(line_streamer)
 
-        # Since we determine `self.parse_metric_families` dynamically from the response and that's done as a
-        # side effect inside the `line_streamer` generator, we need to consume the first line in order to
-        # trigger that side effect.
-        try:
-            line_streamer = chain([next(line_streamer)], line_streamer)
-        except StopIteration:
-            # If line_streamer is an empty iterator, next(line_streamer) fails.
-            return
-
         for metric in self.parse_metric_families(line_streamer):
             self.submit_telemetry_number_of_total_metric_samples(metric)
 
@@ -302,19 +289,6 @@ class OpenMetricsScraper:
                 metric.name = metric.name[len(self.raw_metric_prefix) :]
 
             yield metric
-
-    @property
-    def parse_metric_families(self):
-        media_type = self._content_type.split(';')[0]
-        # Setting `use_latest_spec` forces the use of the OpenMetrics format, otherwise
-        # the format will be chosen based on the media type specified in the response's content-header.
-        # The selection is based on what Prometheus does:
-        # https://github.com/prometheus/prometheus/blob/v2.43.0/model/textparse/interface.go#L83-L90
-        return (
-            parse_openmetrics
-            if self._use_latest_spec or media_type == 'application/openmetrics-text'
-            else parse_prometheus
-        )
 
     def generate_sample_data(self, metric):
         """
@@ -369,8 +343,6 @@ class OpenMetricsScraper:
 
         try:
             with self.get_connection() as connection:
-                # Media type will be used to select parser dynamically
-                self._content_type = connection.headers.get('Content-Type', '')
                 for line in connection.iter_lines(decode_unicode=True):
                     yield line
         except ConnectionError as e:
@@ -415,7 +387,6 @@ class OpenMetricsScraper:
                     response.encoding = 'utf-8'
 
                 self.submit_telemetry_endpoint_response_size(response)
-
                 return response
 
     def send_request(self, **kwargs):
